Add protocol version for gas and conservation
