+++ crates/sui-core/src/transaction_input_checker.rs
@@ -9,7 +9,6 @@ use sui_macros::checked_arithmetic;
 use sui_protocol_config::ProtocolConfig;
 use sui_types::base_types::ObjectRef;
 use sui_types::error::{UserInputError, UserInputResult};
-use sui_types::gas::SuiCostTable;
 use sui_types::messages::{
     TransactionKind, VerifiedExecutableTransaction, VersionedProtocolMessage,
 };
@@ -17,7 +16,7 @@ use sui_types::{
     base_types::{SequenceNumber, SuiAddress},
     error::SuiResult,
     fp_ensure,
-    gas::{self, SuiGasStatus},
+    gas::{SuiCostTable, SuiGasStatus},
     messages::{InputObjectKind, InputObjects, TransactionData, TransactionDataAPI},
     object::{Object, Owner},
 };
@@ -173,8 +172,9 @@ async fn check_gas(
     gas_price: u64,
     tx_kind: &TransactionKind,
 ) -> SuiResult<SuiGasStatus<'static>> {
+    let protocol_config = epoch_store.protocol_config();
     if tx_kind.is_system_tx() {
-        Ok(SuiGasStatus::new_unmetered())
+        Ok(SuiGasStatus::new_unmetered(protocol_config))
     } else {
         // gas price must be bigger or equal to reference gas price
         let reference_gas_price = epoch_store.reference_gas_price();
@@ -205,20 +205,17 @@ async fn check_gas(
         }
 
         // check balance and coins consistency
-        let protocol_config = epoch_store.protocol_config();
-        let cost_table = SuiCostTable::new(protocol_config);
-        gas::check_gas_balance(
-            gas_object,
-            more_gas_objects,
-            gas_budget,
-            gas_price,
-            &cost_table,
-        )?;
-
-        // make the gas status to be used by execution
-        let storage_gas_price = protocol_config.storage_gas_price();
-        gas::start_gas_metering(gas_budget, gas_price, storage_gas_price, cost_table)
-            .map_err(|e| e.into())
+        if protocol_config.gas_model_version() > 1 {
+            panic!("TBD");
+        } else {
+            let cost_table = SuiCostTable::new(protocol_config);
+            cost_table.check_gas_balance(gas_object, more_gas_objects, gas_budget, gas_price)?;
+            Ok(SuiGasStatus::new_with_budget(
+                gas_budget,
+                gas_price,
+                protocol_config,
+            ))
+        }
     }
 }
