+++ crates/sui-core/src/unit_tests/gas_tests.rs
@@ -11,16 +11,16 @@ use move_core_types::ident_str;
 use once_cell::sync::Lazy;
 use sui_framework::make_system_objects;
 use sui_types::crypto::AccountKeyPair;
+use sui_types::gas::{SuiCostTable, SuiGasStatusAPI};
 use sui_types::gas_coin::GasCoin;
 use sui_types::is_system_package;
 use sui_types::object::GAS_VALUE_FOR_TESTING;
 use sui_types::programmable_transaction_builder::ProgrammableTransactionBuilder;
 use sui_types::utils::to_sender_signed_transaction;
-use sui_types::{base_types::dbg_addr, crypto::get_key_pair, gas::SuiGasStatus};
+use sui_types::{base_types::dbg_addr, crypto::get_key_pair};
 
-static MAX_GAS_BUDGET: Lazy<u64> = Lazy::new(|| SuiCostTable::new_for_testing().max_gas_budget);
-static MIN_GAS_BUDGET: Lazy<u64> =
-    Lazy::new(|| SuiCostTable::new_for_testing().min_gas_budget_external());
+static MAX_GAS_BUDGET: Lazy<u64> = Lazy::new(|| SuiCostTable::new_for_testing().max_gas_budget());
+static MIN_GAS_BUDGET: Lazy<u64> = Lazy::new(|| SuiCostTable::new_for_testing().min_gas_budget());
 
 #[tokio::test]
 async fn test_tx_less_than_minimum_gas_budget() {
@@ -108,19 +108,15 @@ async fn test_native_transfer_sufficient_gas() -> SuiResult {
 
     // Mimic the process of gas charging, to check that we are charging
     // exactly what we should be charging.
-    let mut gas_status = SuiGasStatus::new_with_budget(
-        *MAX_GAS_BUDGET,
-        1.into(),
-        1.into(),
-        SuiCostTable::new_for_testing(),
-    );
+    let mut gas_status =
+        SuiCostTable::new_for_testing().into_gas_status_for_testing(*MAX_GAS_BUDGET, 1, 1);
     let obj_size = object.object_size_for_gas_metering();
     let gas_size = gas_object.object_size_for_gas_metering();
 
     gas_status.charge_storage_read(obj_size + gas_size)?;
     gas_status.bucketize_computation()?;
-    gas_status.charge_storage_mutation(obj_size, 0.into())?;
-    gas_status.charge_storage_mutation(gas_size, 0.into())?;
+    gas_status.charge_storage_mutation(obj_size, 0)?;
+    gas_status.charge_storage_mutation(gas_size, 0)?;
     let summary = gas_status.summary();
     assert_eq!(gas_cost, &summary);
     Ok(())
@@ -262,7 +258,6 @@ async fn test_native_transfer_insufficient_gas_execution() {
     );
 }
 
-// disabled because it violates the SUI conservation checks
 #[tokio::test]
 async fn test_publish_gas() -> anyhow::Result<()> {
     let (sender, sender_key): (_, AccountKeyPair) = get_key_pair();
@@ -304,12 +299,8 @@ async fn test_publish_gas() -> anyhow::Result<()> {
     };
 
     // Mimic the gas charge behavior and cross check the result with above.
-    let mut gas_status = SuiGasStatus::new_with_budget(
-        *MAX_GAS_BUDGET,
-        1.into(),
-        1.into(),
-        SuiCostTable::new_for_testing(),
-    );
+    let mut gas_status =
+        SuiCostTable::new_for_testing().into_gas_status_for_testing(*MAX_GAS_BUDGET, 1, 1);
     gas_status.charge_storage_read(
         genesis_objects
             .iter()
@@ -320,8 +311,8 @@ async fn test_publish_gas() -> anyhow::Result<()> {
     )?;
     gas_status.charge_storage_read(gas_object.object_size_for_gas_metering())?;
     gas_status.charge_publish_package(publish_bytes.iter().map(|v| v.len()).sum())?;
-    gas_status.charge_storage_mutation(package.object_size_for_gas_metering(), 0.into())?;
-    gas_status.charge_storage_mutation(gas_object.object_size_for_gas_metering(), 0.into())?;
+    gas_status.charge_storage_mutation(package.object_size_for_gas_metering(), 0)?;
+    gas_status.charge_storage_mutation(gas_object.object_size_for_gas_metering(), 0)?;
     // Actual gas cost will be greater than the expected summary because of the cost to discard the
     // `UpgradeCap`.
     let gas_summary = gas_status.summary();
@@ -410,12 +401,8 @@ async fn test_move_call_gas() -> SuiResult {
     // Mimic the gas charge behavior and cross check the result with above. Do not include
     // computation cost calculation as it would require hard-coding a constant representing VM
     // execution cost which is quite fragile.
-    let mut gas_status = SuiGasStatus::new_with_budget(
-        GAS_VALUE_FOR_TESTING,
-        1.into(),
-        1.into(),
-        SuiCostTable::new_for_testing(),
-    );
+    let mut gas_status =
+        SuiCostTable::new_for_testing().into_gas_status_for_testing(*MAX_GAS_BUDGET, 1, 1);
     let package_object = authority_state
         .get_object(&package_object_ref.0)
         .await?
@@ -427,8 +414,8 @@ async fn test_move_call_gas() -> SuiResult {
         .get_object(&effects.created()[0].0 .0)
         .await?
         .unwrap();
-    gas_status.charge_storage_mutation(created_object.object_size_for_gas_metering(), 0.into())?;
-    gas_status.charge_storage_mutation(gas_object.object_size_for_gas_metering(), 0.into())?;
+    gas_status.charge_storage_mutation(created_object.object_size_for_gas_metering(), 0)?;
+    gas_status.charge_storage_mutation(gas_object.object_size_for_gas_metering(), 0)?;
     let new_cost = gas_status.summary();
     assert_eq!(gas_cost.storage_cost, new_cost.storage_cost);
     // This is the total amount of storage cost paid. We will use this
@@ -465,21 +452,13 @@ async fn test_move_call_gas() -> SuiResult {
 
 #[tokio::test]
 async fn test_storage_gas_unit_price() -> SuiResult {
-    let mut gas_status1 = SuiGasStatus::new_with_budget(
-        *MAX_GAS_BUDGET,
-        1.into(),
-        1.into(),
-        SuiCostTable::new_for_testing(),
-    );
-    gas_status1.charge_storage_mutation(200, 5.into())?;
+    let mut gas_status1 =
+        SuiCostTable::new_for_testing().into_gas_status_for_testing(*MAX_GAS_BUDGET, 1, 1);
+    gas_status1.charge_storage_mutation(200, 5)?;
     let gas_cost1 = gas_status1.summary();
-    let mut gas_status2 = SuiGasStatus::new_with_budget(
-        *MAX_GAS_BUDGET,
-        1.into(),
-        3.into(),
-        SuiCostTable::new_for_testing(),
-    );
-    gas_status2.charge_storage_mutation(200, 5.into())?;
+    let mut gas_status2 =
+        SuiCostTable::new_for_testing().into_gas_status_for_testing(*MAX_GAS_BUDGET, 1, 3);
+    gas_status2.charge_storage_mutation(200, 5)?;
     let gas_cost2 = gas_status2.summary();
     // Computation unit price is the same, hence computation cost should be the same.
     assert_eq!(gas_cost1.computation_cost, gas_cost2.computation_cost);
