+++ crates/sui-types/src/lib.rs
@@ -35,6 +35,7 @@ pub mod dynamic_field;
 pub mod event;
 pub mod gas;
 pub mod gas_coin;
+pub mod gas_model;
 pub mod governance;
 pub mod id;
 pub mod in_memory_storage;
